AI-IntakeandAssessment/Questionnaire_agent/func_indexer/requirements.txt
azure-functions
azure-search-documents
azure-storage-blob
azure-identity
python-docx
pypdf
pandas
openpyxl
numpy
tenacity
